backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers

# Automated KYC
pytesseract
opencv-contrib-python
pillow
numpy